Initial implementation of merged comment cells
This is rather filthy code. To do this, we:
- Disable the Status, Proposal, Run, Timestamp, and Comment columns from
  moving. Only the variable columns can be moved now.
- Set column spans from the first column up till the timestamp, and from the
  comment column till infinity (actually 1000 but that's plenty).
- Clear and recalculate the spans when the sorting changes and rows are
  inserted.
- Make the model only return a CheckState if not on a comment row.

The reason first five columns need to be fixed in position is because allowing
them to move makes calculating the column spans difficult. Naively, one would
think the following algorithm would work:
- If the comment column is after the timestamp column:
  - Span the columns from 0 to the timestamp.
  - Span any columns between the timestamp and the comment (care needs to be
    taken here, span from timestamp_col + 1 to keep the cells visually
    separate).
  - Span all columns after the comment.
- Similarly if the comment column is before the timestamp column.

However, moving columns is an entirely visual change. QHeaderView supports both
logical and visual indices and QTableView.setSpan() works with logical
indices (as far as I can tell). Which means that figuring out which logical
indices to span from the visual indices is rather tricky.

Alternatively, we could just reorder the columns in the DataFrame whenever a
column is moved and then calculate the spans as described above. I think in the
long term this is the best option.