Fix bug with detecting changing variables
Previously `variable_is_constant()` only took in the table index as an argument,
which would be cached by @lru_cache. However if the table sorting is changed,
the index previously used for one row may point to a different row and the wrong
value might be returned from the cache for a given run. Now
`variable_is_constant()` is parameterized by the run, proposal, and quantity to
be sure to return the right value.

Note: need to move this logic into the backend.