Fix the instructions for connecting to XFELs broker offline
I'm not sure how this worked originally, but when testing it I found that the
KafkaConsumer will connect to the bootstrap server (localhost) correctly with
the AMORE_BROKER variable set, but it will then try to connect to
exflwebstor01.desy.de and hang. I'm guessing that when it connects to the
bootstrap server it gets a list of all brokers and then tries to connect to
them; and since it sees localhost and exflwebstor01.desy.de as different brokers
it'll try to connect to exflwebstor01.desy.de too.