Reprocesses runs from the GUI
First prototype of the feature. From the tableview a context menu is available.
From the menu the user can request to reprocess the selected runs.
The seleceted runs are added to a Queue. Such queue is being watched in a QThread.
When non-empty the runs are, one by one, reprocessed by the worker inside the thread.