Supports reprocessing from the GUI
A copy of my old branch with a few alterations. Enables reprocessing
from the GUI by right-clicking a given row (or selection of) on the run
table. A color scheme indicates the reprocessing status. Support for
SLURM jobs it's still not availiable, thus the color scheme might be
missleading if SLURM jobs are required for some variables.