Adds jump to run feature
At the top right corner of the GUI a magnifing glass icon is added.
Upon clinking on it, a QLineEdit widget is poped. Upon entering a valid
run number and pressing the return key the tableview scrolls to the
corresponding row. When new rows are added, automatic scrolling is also
supported.