Skip non-existent runs when reprocessing
Previously these would be attempted anyway, which has some disadvantages:
- Reprocessing with e.g. `amore-proto reprocess $(seq 1 100)` would fail if any
  run in-between was missing, so it could take multiple invocations to reprocess
  a range.
- Even if a run didn't exist a database row would still be created for it, which
  meant that any future use of `amore-proto reprocess all` would also crash when
  it got to that run.

Note that this doesn't forbid including non-existent runs in the table, it just
skips them when reprocessing. That's important because at some point when we
hook up the DAQ as a trigger we'll record *all* created runs, not just those
that have been migrated.