Always evaluate the context file in a separate process
The GUI used to have a special case for databases using the default environment
where it would evaluate the context file in the same process. This is bad
because then the context file might not be evaluated in the database directory,
which would break local imports. Now we always evaluate it with
`get_context_file()` in a separate process in the database directory.